Readme file, project report and screenshots are also being provided.
The project repository contain all the necessary files, screenshots of final results on different inputs of messages and working description of project in readme file.
One time padding is one of the most secure method of encryption being used in transaction. This project is to based on analysis of the code and working on its implementation with few functions being described to analyze the code more in depth.